+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
-
+## 0.0.6
+- Fixed the error handling in campusCollege.js
+- Fixed the error handling in cashier.js
+- Fixed the error handling in college.js
+- Fixed the error handling in collegeCourses.js
+- Fixed the error handling in course.js
+- Fixed the error handling in document.js
+- Fixed the error handling in mainCampus.js
 ## 0.0.5
 
 - Added curriculum model